+++ docs/conf.py
@@ -39,7 +39,7 @@ copyright = '2015-2022, Adam Mansur'
 author = 'Adam Mansur'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0'
+release = '1.1'
 
 
 # -- General configuration ---------------------------------------------------