+++ stitch2d/__init__.py
@@ -3,6 +3,6 @@ from .mosaic import Mosaic, StructuredMosaic, create_mosaic, build_grid, is_grid
 from .tile import Tile, OpenCVTile, ScikitImageTile
 
 
-__version__ = "1.0"
+__version__ = "1.1"
 __author__ = "Adam Mansur"
 __credits__ = "Smithsonian National Museum of Natural History"