+++ CHANGELOG.md
@@ -8,12 +8,20 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+1.1
+---
+
+### Added
+
+-   Added optional channel_order kwarg to the stitch method. Allows
+    users to specify the channel order in the stitched array.
+
 1.0
 ---
 
 ### Added
 
--   Added scikit-image backend in addition to opencv (but it's very
+-   Added scikit-image backend in addition to opencv (but it’s very
     slow)
 -   Added tests